chore: Move bike insights models to models/bike_insights and update imports
- Relocated agent.py and bikes.py to models/bike_insights directory
- Updated import paths in bike_insights.py to reflect new location
