+++ ingenious/ingenious_extensions_template/models/bike_insights/bikes.py
